+++ src/main/java/qna/domain/Answers.java
@@ -1,2 +1,44 @@
-package qna.domain;public class Answers {
+package qna.domain;
+
+
+import org.hibernate.annotations.Where;
+import qna.CannotDeleteException;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+import javax.persistence.OrderBy;
+import java.util.ArrayList;
+import java.util.List;
+
+@Embeddable
+public class Answers {
+
+    @OneToMany(mappedBy = "question", cascade = CascadeType.ALL)
+    @Where(clause = "deleted = false")
+    @OrderBy("id ASC")
+    private List<Answer> answers = new ArrayList<>();
+
+    public Answers() {
+    }
+
+    public Answers(List<Answer> answers) {
+        this.answers = answers;
+    }
+
+    public List<DeleteHistory> delete(User loginUser) throws CannotDeleteException {
+        List<DeleteHistory> deleteHistories = new ArrayList<>();
+        for (Answer answer : answers) {
+            deleteHistories.add(answer.delete(loginUser));
+        }
+        return deleteHistories;
+    }
+
+    public List<Answer> getAnswers() {
+        return new ArrayList<>(answers);
+    }
+
+    public void add(Answer answer) {
+        answers.add(answer);
+    }
 }
